Bug 2063033 - Do not delete borrowed MozFramebuffer color backings. r=gfx-reviewers,stransky,lsalzman
Differential Revision: https://phabricator.services.mozilla.com/D318191

+++ gfx/gl/MozFramebuffer.cpp
@@ -69,37 +69,39 @@ UniquePtr<MozFramebuffer> MozFramebuffer::Create(GLContext* const gl,
                     (depth || stencil) ? DepthAndStencilBuffer::Create(
                                              gl, size, samples, depth, stencil)
                                        : nullptr,
-                    colorTarget, colorName);
+                    colorTarget, colorName, ColorBackingOwnership::Owned);
 }
 
 UniquePtr<MozFramebuffer> MozFramebuffer::CreateForBacking(
     GLContext* const gl, const gfx::IntSize& size, const uint32_t samples,
-    bool depth, bool stencil, const GLenum colorTarget,
-    const GLuint colorName) {
+    bool depth, bool stencil, const GLenum colorTarget, const GLuint colorName,
+    ColorBackingOwnership colorBackingOwnership) {
   return CreateImpl(gl, size, samples,
                     (depth || stencil) ? DepthAndStencilBuffer::Create(
                                              gl, size, samples, depth, stencil)
                                        : nullptr,
-                    colorTarget, colorName);
+                    colorTarget, colorName, colorBackingOwnership);
 }
 
 /* static */ UniquePtr<MozFramebuffer>
 MozFramebuffer::CreateForBackingWithSharedDepthAndStencil(
     const gfx::IntSize& size, const uint32_t samples, GLenum colorTarget,
     GLuint colorName,
-    const RefPtr<DepthAndStencilBuffer>& depthAndStencilBuffer) {
+    const RefPtr<DepthAndStencilBuffer>& depthAndStencilBuffer,
+    ColorBackingOwnership colorBackingOwnership) {
   auto gl = depthAndStencilBuffer->gl();
   if (!gl || !gl->MakeCurrent()) {
     return nullptr;
   }
   return CreateImpl(gl, size, samples, depthAndStencilBuffer, colorTarget,
-                    colorName);
+                    colorName, colorBackingOwnership);
 }
 
 /* static */ UniquePtr<MozFramebuffer> MozFramebuffer::CreateImpl(
     GLContext* const gl, const gfx::IntSize& size, const uint32_t samples,
     const RefPtr<DepthAndStencilBuffer>& depthAndStencilBuffer,
-    const GLenum colorTarget, const GLuint colorName) {
+    const GLenum colorTarget, const GLuint colorName,
+    ColorBackingOwnership colorBackingOwnership) {
   GLuint fb = gl->CreateFramebuffer();
   const ScopedBindFramebuffer bindFB(gl, fb);
 
@@ -136,8 +138,9 @@ MozFramebuffer::CreateForBackingWithSharedDepthAndStencil(
     return nullptr;
   }
 
-  return UniquePtr<MozFramebuffer>(new MozFramebuffer(
-      gl, size, fb, samples, depthAndStencilBuffer, colorTarget, colorName));
+  return UniquePtr<MozFramebuffer>(
+      new MozFramebuffer(gl, size, fb, samples, depthAndStencilBuffer,
+                         colorTarget, colorName, colorBackingOwnership));
 }
 
 /* static */ RefPtr<DepthAndStencilBuffer> DepthAndStencilBuffer::Create(
@@ -187,14 +190,16 @@ MozFramebuffer::CreateForBackingWithSharedDepthAndStencil(
 MozFramebuffer::MozFramebuffer(
     GLContext* const gl, const gfx::IntSize& size, GLuint fb,
     const uint32_t samples, RefPtr<DepthAndStencilBuffer> depthAndStencilBuffer,
-    const GLenum colorTarget, const GLuint colorName)
+    const GLenum colorTarget, const GLuint colorName,
+    ColorBackingOwnership colorBackingOwnership)
     : mWeakGL(gl),
       mSize(size),
       mSamples(samples),
       mFB(fb),
       mColorTarget(colorTarget),
       mDepthAndStencilBuffer(std::move(depthAndStencilBuffer)),
-      mColorName(colorName) {
+      mColorName(colorName),
+      mColorBackingOwnership(colorBackingOwnership) {
   MOZ_ASSERT(mColorTarget);
   MOZ_ASSERT(mColorName);
 }
@@ -207,7 +212,9 @@ MozFramebuffer::~MozFramebuffer() {
 
   gl->DeleteFramebuffer(mFB);
 
-  DeleteByTarget(gl, mColorTarget, mColorName);
+  if (mColorBackingOwnership == ColorBackingOwnership::Owned) {
+    DeleteByTarget(gl, mColorTarget, mColorName);
+  }
 }
 
 bool MozFramebuffer::HasDepth() const {
